backend/AI_Features/foundational-essays/app/requirements.txt
Flask==2.1.0
google-generativeai==0.3.1
python-dotenv==0.19.2